Factor out drawn_sprite to its own file
Also add an utility function to calculate the bounding box of a sprite vector.
